+++ packages/boilerplate/template/src/shared/middleware.ts
@@ -1,10 +1,16 @@
 
 import { createGetTables, createGetUser, createGetUserTeams } from "@calatrava/middleware";
-import { TableKeyManager } from "@calatrava/datawrapper";
+import { TableKeyManager, DBKeys } from "@calatrava/datawrapper";
 import { Tables, User, Team, Teammate } from "./types";
 import { decodeToken } from "./token";
 
-const tableKeyManager = new TableKeyManager();
+const tableKeysMap = {
+  [Tables.Users]: {
+    [DBKeys.Partition]: ["userId"],
+  },
+};
+
+const tableKeyManager = new TableKeyManager(tableKeysMap, Tables);
 
 export const getTables = createGetTables(tableKeyManager);
 export const getUser = createGetUser<User>(Tables.Users, decodeToken);